+++ CHANGELOG.markdown
@@ -2,6 +2,11 @@
 
 ## Next release
 * Make middleware compatible with both `MIDDLEWARE` and `MIDDLEWARE_CLASSES`.
+* Ensure all `ForeignKey`s have an `on_delete` explicitly specified for Django 2.2 compatibility.
+* Rename "Search engine optimization" fieldset of SearchMetaBaseAdmin to "SEO".
+
+## 4.1.0 - 25/10/2019
+* Make it possible for ContentBase derivatives to override how they are searched by Watson.
 
 ## 4.0.10 - 18/10/2019
 * Fix `Video.embed_html()` function using the wrong renderer for local mp4 files.